+++ NEWS.md
@@ -1,4 +1,19 @@
 
+# parlitools 0.3.1.9000
+
+* Changed `gss_code` to `ons_const_id` in the `leave_votes_west` dataset
+  to ensure consistent naming of parliamentary constituency codes
+  
+* Added Northern Ireland GE results
+
+
+# parlitools 0.3.1
+
+* Fixed formatting of percentages in the `leave_votes_west` dataset so they are 
+  all scaled on 0-1 rather than some on 0-1 and some on 0-100.
+  
+* Improved the organisation of the British Election Study dataset documentation.
+
 
 # parlitools 0.3.0